backend/requirements.txt
fastapi
uvicorn
supabase
pinecone
openai
pydantic
python-dotenv
python-multipart
langchain
langchain-openai
langchain-community
langgraph
PyPDF2
python-jose[cryptography]
passlib[bcrypt]
composio-langchain
cohere
